requirements.txt
fastapi
uvicorn
yt-dlp
mutagen
pillow
requests
python-multipart
xmltodict